+++ src/KingmakerDiceRoller/CharacterCreation/IRollUiCommandTarget.cs
@@ -6,6 +6,7 @@ namespace KingmakerDiceRoller.CharacterCreation
     {
         RollUiSnapshot UiSnapshot { get; }
         RollSession ActiveSession { get; }
+        bool CanAttachNativePanel { get; }
         bool TryRoll(out string error);
         bool TryReroll(out string error);
         bool TryMoveAssignment(AbilityScore ability, bool moveUp, out string error);