+++ src/KingmakerDiceRoller/Properties/AssemblyInfo.cs
@@ -8,5 +8,6 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyCopyright("Copyright © 2026 Howie Reith")]
 [assembly: ComVisible(false)]
 [assembly: Guid("f2fb5a20-7160-4cc2-b790-d84229b894bd")]
-[assembly: AssemblyVersion("0.0.1.0")]
-[assembly: AssemblyFileVersion("0.0.1.0")]
+[assembly: AssemblyVersion("0.1.0.0")]
+[assembly: AssemblyFileVersion("0.1.0.0")]
+[assembly: AssemblyInformationalVersion("0.1.0-alpha.1")]